Make card payloads compatible with ember-mobiledoc-editor
ember-mobiledoc-editor passes the payload to card components as `data`, but ember-mobiledoc-dom-renderer passes the payload as `payload`. This makes it awkward to use the same set of display card components in both environments.

This changes standardizes on `data`, mostly because it meant changing the simpler side. But I think `payload` is a better name, and we could also change the editor to that instead.